+++ source/marketparticipant/Energinet.DataHub.MarketParticipant.EntryPoint.WebApi/MarketParticipantWebApiModuleExtensions.cs
@@ -16,7 +16,6 @@ using Azure.Identity;
 using Azure.Security.KeyVault.Keys;
 using Azure.Security.KeyVault.Secrets;
 using Energinet.DataHub.MarketParticipant.Application.Services;
-using Energinet.DataHub.MarketParticipant.Authorization.Services;
 using Energinet.DataHub.MarketParticipant.Common;
 using Energinet.DataHub.MarketParticipant.Common.Options;
 using Energinet.DataHub.MarketParticipant.EntryPoint.WebApi.Options;
@@ -62,15 +61,6 @@ public static class MarketParticipantWebApiModuleExtensions
             var clock = provider.GetRequiredService<IClock>();
             var keyClient = new KeyClient(options.Value.TokenSignKeyVault, tokenCredentials);
             return new SigningKeyRing(clock, keyClient, options.Value.TokenSignKeyName);
-        });
-
-        _ = services.AddSingleton<IAuthorizationService>(provider =>
-        {
-            var tokenCredentials = new DefaultAzureCredential();
-            var options = provider.GetRequiredService<IOptions<KeyVaultOptions>>();
-            var keyClient = new KeyClient(options.Value.TokenSignKeyVault, tokenCredentials);
-            var logger = provider.GetRequiredService<ILogger<AuthorizationService>>();
-            return new AuthorizationService(options.Value.TokenSignKeyVault, options.Value.AuthSignKeyName, logger);
         });
 
         services.AddSingleton(provider =>