+++ docs/Energinet.DataHub.MarketParticipant.Authorization/release-notes/release-notes.md
@@ -1,5 +1,17 @@
 # Energinet.DataHub.MarketParticipant.Authorization release notes
 
+## Version 1.0.2
+
+- Added energysupplier role validation for access validation.
+
+## Version 1.0.1
+
+- Added support for key rotation.
+
+## Version 1.0.0
+
+- First implementation of Authorization
+
 ## Version 0.1.1
 
 - Placeholder release.